refactor core projects
