--- application/src/main/java/com/pharmacy/application/exceptions/errors/DateExpiredException.java
@@ -1,7 +1,0 @@
-package com.pharmacy.application.exceptions.errors;
-
-public class DateExpiredException extends RuntimeException {
-    public DateExpiredException(String message) {
-        super(message);
-    }
-}
