--- application/src/main/java/com/pharmacy/application/exceptions/errors/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.pharmacy.application.exceptions.errors;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
