--- application/src/main/java/com/pharmacy/application/exceptions/errors/WrongManufacturerException.java
@@ -1,7 +1,0 @@
-package com.pharmacy.application.exceptions.errors;
-
-public class WrongManufacturerException extends RuntimeException {
-    public WrongManufacturerException(String message) {
-        super(message);
-    }
-}
