+++ core/application-contracts/.gitignore
