+++ core/application-contracts/src/main/java/com/pharmacy/application/contracts/dtos/CreateProductDTO.java
@@ -1,0 +1,6 @@
+package com.pharmacy.application.contracts.dtos;
+
+import java.util.Date;
+
+public record CreateProductDTO(String name, ManufacturerDTO manufacturer, int price, Date expiryDate) {
+}
