+++ core/application-contracts/src/main/java/com/pharmacy/application/contracts/dtos/ManufacturerProductsCountDTO.java
