+++ core/application-contracts/src/main/java/com/pharmacy/application/contracts/dtos/UpdateProductDTO.java
@@ -1,0 +1,6 @@
+package com.pharmacy.application.contracts.dtos;
+
+import java.util.Date;
+
+public record UpdateProductDTO(String name, ManufacturerDTO manufacturer, int price, Date expiryDate) {
+}
