+++ core/application-contracts/src/main/java/com/pharmacy/application/contracts/mappers/ProductMapper.java
@@ -1,0 +1,13 @@
+package com.pharmacy.application.contracts.mappers;
+
+import com.pharmacy.application.contracts.dtos.ProductDTO;
+import com.pharmacy.domain.model.Product;
+
+import java.util.List;
+
+public interface ProductMapper {
+
+    ProductDTO toDto(Product product);
+
+    List<ProductDTO> toDto(List<Product> productList);
+}
