+++ core/application-contracts/src/main/java/com/pharmacy/application/contracts/services/ManufacturerAppService.java
