+++ core/application/.gitignore
