+++ core/domain/src/main/java/com/pharmacy/domain/exceptions/CustomArgumentException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.domain.exceptions;
+
+public class CustomArgumentException extends RuntimeException {
+    public CustomArgumentException(String message) {
+        super(message);
+    }
+}
