+++ core/domain/src/main/java/com/pharmacy/domain/exceptions/ProductDateExpiredException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.domain.exceptions;
+
+public class ProductDateExpiredException extends RuntimeException {
+    public ProductDateExpiredException(String message) {
+        super(message);
+    }
+}
