+++ core/domain/src/main/java/com/pharmacy/domain/exceptions/ProductExistsException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.domain.exceptions;
+
+public class ProductExistsException extends RuntimeException {
+    public ProductExistsException(String message) {
+        super(message);
+    }
+}
