+++ core/domain/src/main/java/com/pharmacy/domain/exceptions/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.domain.exceptions;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+}
