+++ core/domain/src/main/java/com/pharmacy/domain/exceptions/WrongManufacturerException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.domain.exceptions;
+
+public class WrongManufacturerException extends RuntimeException {
+    public WrongManufacturerException(String message) {
+        super(message);
+    }
+}
