+++ core/domain/src/main/java/com/pharmacy/domain/model/Default.java
