+++ core/domain/src/main/java/com/pharmacy/domain/repositories/ManufacturerRepository.java
@@ -1,0 +1,20 @@
+package com.pharmacy.domain.repositories;
+
+import com.pharmacy.domain.model.Manufacturer;
+import com.pharmacy.domain.model.Product;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface ManufacturerRepository {
+    List<Manufacturer> findAll();
+
+    Optional<Manufacturer> findById(UUID id);
+
+    boolean existsById(UUID id);
+//
+//    List<CountManuProdsDTO> countManufacturersProducts();
+//
+//    List<Manufacturer> getManufacturersWithProducts();
+}
