--- domain/src/main/java/com/pharmacy/domain/repositories/ManufacturerRepository.java
@@ -1,16 +1,0 @@
-package com.pharmacy.domain.repositories;
-
-import com.pharmacy.domain.model.Manufacturer;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface ManufacturerRepository {
-    List<Manufacturer> findAll();
-
-    boolean existsById(UUID id);
-//
-//    List<CountManuProdsDTO> countManufacturersProducts();
-//
-//    List<Manufacturer> getManufacturersWithProducts();
-}
